+++ lib/src/generator/enumclass_generator.dart
@@ -87,7 +87,8 @@ extension EnumType2EnumClass on EnumType {
   }
 
   String _enumWithoutAnnotations(SquintGeneratorOptions options) =>
-      _generateEnumerationClass(className: className, members: values);
+      _generateEnumerationClass(
+          className: className, members: values..add(_noneValue));
 
   String _enumWithAnnotations(SquintGeneratorOptions options) {
     final members = <String>[];
@@ -98,9 +99,12 @@ extension EnumType2EnumClass on EnumType {
       members.add("""@JsonValue("${valuesJSON[index]}") ${values[index]}""");
       index += 1;
     }
-
+    members.add("""@JsonValue("") $_noneValue""");
     return _generateEnumerationClass(className: className, members: members);
   }
+
+  String get _noneValue =>
+      values.every((e) => e == e.toUpperCase()) ? "NONE" : "none";
 
   String _generateEnumerationClass({
     required String className,